Restructure namespace for codemagic.tools.app_store_connect
