+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "codemagic-cli-tools"
-version = "0.48.2"
+version = "0.49.0"
 description = "CLI tools used in Codemagic builds"
 readme = "README.md"
 authors = [
