+++ src/codemagic/tools/app_store_connect/__init__.py
@@ -1,0 +1,1 @@
+from .app_store_connect import AppStoreConnect
