+++ src/codemagic/tools/app_store_connect/__main__.py
@@ -1,0 +1,4 @@
+if __name__ == "__main__":
+    from .app_store_connect import AppStoreConnect
+
+    AppStoreConnect.invoke_cli()
