+++ src/codemagic/tools/app_store_connect/abstract_base_action.py
@@ -1,52 +1,74 @@
 from __future__ import annotations
 
-import logging
-import pathlib
 from abc import ABCMeta
 from abc import abstractmethod
-from datetime import datetime
-from typing import List
-from typing import Optional
-from typing import Sequence
-from typing import Tuple
-from typing import Union
+from typing import TYPE_CHECKING
 
-from codemagic.apple import AppStoreConnectApiClient
-from codemagic.apple.app_store_connect import IssuerId
-from codemagic.apple.app_store_connect import KeyIdentifier
-from codemagic.apple.resources import App
-from codemagic.apple.resources import AppStoreState
-from codemagic.apple.resources import AppStoreVersion
-from codemagic.apple.resources import AppStoreVersionLocalization
-from codemagic.apple.resources import AppStoreVersionSubmission
-from codemagic.apple.resources import BetaAppReviewSubmission
-from codemagic.apple.resources import BetaBuildLocalization
-from codemagic.apple.resources import BetaReviewState
-from codemagic.apple.resources import Build
-from codemagic.apple.resources import BuildProcessingState
-from codemagic.apple.resources import Locale
+from codemagic.apple.resources import BundleIdPlatform
+from codemagic.apple.resources import CertificateType
 from codemagic.apple.resources import Platform
-from codemagic.apple.resources import ReleaseType
-from codemagic.apple.resources import ResourceId
-from codemagic.apple.resources import ReviewSubmission
-from codemagic.apple.resources import ReviewSubmissionItem
-from codemagic.apple.resources import ReviewSubmissionState
+from codemagic.apple.resources import ProfileType
 from codemagic.mixins import PathFinderMixin
 
-from .arguments import AppStoreVersionInfo
-from .arguments import AppStoreVersionLocalizationInfo
-from .arguments import BetaBuildInfo
-from .arguments import Types
-from .resource_manager_mixin import ResourceManagerMixin
-from .resource_printer import ResourcePrinter
+from .mixins import ResourceManagerMixin
+from .mixins import SigningFileSaverMixin
 
-AppStoreVersionLocalizationInfos = Union[
-    List[AppStoreVersionLocalizationInfo],
-    Types.AppStoreVersionLocalizationInfoArgument,
-]
+if TYPE_CHECKING:
+    import logging
+    import pathlib
+    from datetime import datetime
+    from typing import List
+    from typing import Optional
+    from typing import Sequence
+    from typing import Tuple
+    from typing import Union
+
+    from codemagic.apple import AppStoreConnectApiClient
+    from codemagic.apple.app_store_connect import IssuerId
+    from codemagic.apple.app_store_connect import KeyIdentifier
+    from codemagic.apple.resources import App
+    from codemagic.apple.resources import AppStoreState
+    from codemagic.apple.resources import AppStoreVersion
+    from codemagic.apple.resources import AppStoreVersionLocalization
+    from codemagic.apple.resources import AppStoreVersionSubmission
+    from codemagic.apple.resources import BetaAppReviewSubmission
+    from codemagic.apple.resources import BetaBuildLocalization
+    from codemagic.apple.resources import BetaReviewState
+    from codemagic.apple.resources import Build
+    from codemagic.apple.resources import BuildProcessingState
+    from codemagic.apple.resources import BundleId
+    from codemagic.apple.resources import Device
+    from codemagic.apple.resources import DeviceStatus
+    from codemagic.apple.resources import Locale
+    from codemagic.apple.resources import Profile
+    from codemagic.apple.resources import ProfileState
+    from codemagic.apple.resources import ReleaseType
+    from codemagic.apple.resources import ResourceId
+    from codemagic.apple.resources import ReviewSubmission
+    from codemagic.apple.resources import ReviewSubmissionItem
+    from codemagic.apple.resources import ReviewSubmissionState
+    from codemagic.apple.resources import SigningCertificate
+
+    from ...models import PrivateKey
+    from .arguments import AppStoreVersionInfo
+    from .arguments import AppStoreVersionLocalizationInfo
+    from .arguments import BetaBuildInfo
+    from .arguments import CertificateArgument
+    from .arguments import Types
+    from .resource_printer import ResourcePrinter
+
+    AppStoreVersionLocalizationInfos = Union[
+        List[AppStoreVersionLocalizationInfo],
+        Types.AppStoreVersionLocalizationInfoArgument,
+    ]
 
 
-class AbstractBaseAction(ResourceManagerMixin, PathFinderMixin, metaclass=ABCMeta):
+class AbstractBaseAction(
+    ResourceManagerMixin,
+    SigningFileSaverMixin,
+    PathFinderMixin,
+    metaclass=ABCMeta,
+):
     logger: logging.Logger
     profiles_directory: pathlib.Path
     certificates_directory: pathlib.Path
@@ -54,6 +76,21 @@ class AbstractBaseAction(ResourceManagerMixin, PathFinderMixin, metaclass=ABCMet
     _key_identifier: Optional[KeyIdentifier]
     _issuer_id: Optional[IssuerId]
     _private_key: Optional[str]
+
+    @staticmethod
+    def _get_certificate_key(
+        certificate_key: Optional[Union[PrivateKey, Types.CertificateKeyArgument]] = None,
+        certificate_key_password: Optional[Types.CertificateKeyPasswordArgument] = None,
+    ) -> Optional[PrivateKey]:
+        if isinstance(certificate_key, PrivateKey):
+            return certificate_key
+        password = certificate_key_password.value if certificate_key_password else None
+        if certificate_key is not None:
+            try:
+                return PrivateKey.from_buffer(certificate_key.value, password)
+            except ValueError:
+                CertificateArgument.PRIVATE_KEY.raise_argument_error("Not a valid certificate private key")
+        return None
 
     # Define signatures for self-reference to other action groups
 
@@ -187,6 +224,51 @@ class AbstractBaseAction(ResourceManagerMixin, PathFinderMixin, metaclass=ABCMet
         raise NotImplementedError()
 
     @abstractmethod
+    def create_bundle_id(
+        self,
+        bundle_id_identifier: str,
+        bundle_id_name: Optional[str] = None,
+        platform: BundleIdPlatform = BundleIdPlatform.IOS,
+        should_print: bool = True,
+    ) -> BundleId:
+        from .action_groups import BundleIdsActionGroup
+
+        _ = BundleIdsActionGroup.create_bundle_id  # Implementation
+        raise NotImplementedError()
+
+    @abstractmethod
+    def create_certificate(
+        self,
+        certificate_type: CertificateType = CertificateType.IOS_DEVELOPMENT,
+        certificate_key: Optional[Union[PrivateKey, Types.CertificateKeyArgument]] = None,
+        certificate_key_password: Optional[Types.CertificateKeyPasswordArgument] = None,
+        p12_container_password: str = "",
+        p12_container_save_path: Optional[pathlib.Path] = None,
+        save: bool = False,
+        should_print: bool = True,
+    ) -> SigningCertificate:
+        from .action_groups import CertificatesActionGroup
+
+        _ = CertificatesActionGroup.create_certificate  # Implementation
+        raise NotImplementedError()
+
+    @abstractmethod
+    def create_profile(
+        self,
+        bundle_id_resource_id: ResourceId,
+        certificate_resource_ids: Sequence[ResourceId],
+        device_resource_ids: Optional[Sequence[ResourceId]] = None,
+        profile_type: ProfileType = ProfileType.IOS_APP_DEVELOPMENT,
+        profile_name: Optional[str] = None,
+        save: bool = False,
+        should_print: bool = True,
+    ) -> Profile:
+        from .action_groups import ProfilesActionGroup
+
+        _ = ProfilesActionGroup.create_profile  # Implementation
+        raise NotImplementedError()
+
+    @abstractmethod
     def create_review_submission(
         self,
         application_id: ResourceId,
@@ -261,6 +343,17 @@ class AbstractBaseAction(ResourceManagerMixin, PathFinderMixin, metaclass=ABCMet
         raise NotImplementedError()
 
     @abstractmethod
+    def get_bundle_id(
+        self,
+        bundle_id_resource_id: ResourceId,
+        should_print: bool = True,
+    ) -> BundleId:
+        from .action_groups import BundleIdsActionGroup
+
+        _ = BundleIdsActionGroup.get_bundle_id  # Implementation
+        raise NotImplementedError()
+
+    @abstractmethod
     def list_app_store_version_localizations(
         self,
         app_store_version_id: ResourceId,
@@ -303,9 +396,69 @@ class AbstractBaseAction(ResourceManagerMixin, PathFinderMixin, metaclass=ABCMet
         build_version_number: Optional[int] = None,
         should_print: bool = True,
     ) -> List[Build]:
-        from ..app_store_connect import AppStoreConnect
+        from codemagic.tools.app_store_connect import AppStoreConnect
 
         _ = AppStoreConnect.list_builds  # Implementation
+        raise NotImplementedError()
+
+    @abstractmethod
+    def list_bundle_ids(
+        self,
+        bundle_id_identifier: Optional[str] = None,
+        bundle_id_name: Optional[str] = None,
+        platform: Optional[BundleIdPlatform] = None,
+        bundle_id_identifier_strict_match: bool = False,
+        should_print: bool = True,
+    ) -> List[BundleId]:
+        from .action_groups import BundleIdsActionGroup
+
+        _ = BundleIdsActionGroup.list_bundle_ids  # Implementation
+        raise NotImplementedError()
+
+    @abstractmethod
+    def list_bundle_id_profiles(
+        self,
+        bundle_id_resource_ids: Sequence[ResourceId],
+        profile_type: Optional[ProfileType] = None,
+        profile_state: Optional[ProfileState] = None,
+        profile_name: Optional[str] = None,
+        save: bool = False,
+        should_print: bool = True,
+    ) -> List[Profile]:
+        from .action_groups import BundleIdsActionGroup
+
+        _ = BundleIdsActionGroup.list_bundle_id_profiles  # Implementation
+        raise NotImplementedError()
+
+    @abstractmethod
+    def list_certificates(
+        self,
+        certificate_types: Optional[Union[CertificateType, Sequence[CertificateType]]] = None,
+        profile_type: Optional[ProfileType] = None,
+        display_name: Optional[str] = None,
+        certificate_key: Optional[Union[PrivateKey, Types.CertificateKeyArgument]] = None,
+        certificate_key_password: Optional[Types.CertificateKeyPasswordArgument] = None,
+        p12_container_password: str = "",
+        save: bool = False,
+        should_print: bool = True,
+        **_deprecated_kwargs,
+    ) -> List[SigningCertificate]:
+        from .action_groups import CertificatesActionGroup
+
+        _ = CertificatesActionGroup.list_certificates  # Implementation
+        raise NotImplementedError()
+
+    @abstractmethod
+    def list_devices(
+        self,
+        platform: Optional[BundleIdPlatform] = None,
+        device_name: Optional[str] = None,
+        device_status: Optional[DeviceStatus] = None,
+        should_print: bool = True,
+    ) -> List[Device]:
+        from .action_groups import DevicesActionGroup
+
+        _ = DevicesActionGroup.list_devices  # Implementation
         raise NotImplementedError()
 
     @abstractmethod
