+++ src/codemagic/tools/app_store_connect/action_groups/apps_action_group.py
