+++ src/codemagic/tools/app_store_connect/actions/__init__.py
@@ -1,3 +1,4 @@
+from .fetch_signing_files_action import FetchSigningFilesAction
 from .latest_build_number_actions import GetLatestAppStoreBuildNumberAction
 from .latest_build_number_actions import GetLatestBuildNumberAction
 from .latest_build_number_actions import GetLatestTestflightBuildNumberAction
