+++ src/codemagic/tools/app_store_connect/errors.py
