+++ src/codemagic/tools/app_store_connect/mixins/__init__.py
@@ -1,0 +1,2 @@
+from .resource_manager_mixin import ResourceManagerMixin
+from .signing_file_saver_mixin import SigningFileSaverMixin
