+++ src/codemagic/tools/app_store_connect/resource_printer.py
