+++ tests/tools/app_store_connect/arguments/test_beta_build_localizations.py
@@ -6,8 +6,8 @@ from argparse import ArgumentTypeError
 import pytest
 from codemagic.apple.resources import Locale
 from codemagic.models.enums import ResourceEnumMeta
-from codemagic.tools.app_store_connect import BetaBuildInfo
-from codemagic.tools.app_store_connect import Types
+from codemagic.tools.app_store_connect.arguments import BetaBuildInfo
+from codemagic.tools.app_store_connect.arguments import Types
 
 
 @pytest.fixture()
